function_app_sb_trigger/requirements.txt
azure-functions
azure-storage-blob
azure-cosmos
azure-servicebus
python-dotenv


azure-ai-documentintelligence
azure-identity
pydantic
langchain
langchain_core
langchain_community
langchain_openai

